requirements_complete.txt
# Tiger系统完整依赖列表

# 核心依赖
pyyaml>=6.0
redis>=4.5.0
psutil>=5.9.0
cryptography>=40.0.0
requests>=2.28.0
aiohttp>=3.8.0
asyncio-mqtt>=0.16.0
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.0.2
alembic>=1.10.0

# 数据分析
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# 技术指标
ta>=0.10.0
# ta-lib  # 需要单独安装系统库

# 交易所API
ccxt>=3.0.0
python-binance>=1.0.17

# 区块链
web3>=6.0.0
eth-account>=0.8.0

# 社交媒体
tweepy>=4.13.0
praw>=7.7.0  # Reddit

# 消息通知
python-telegram-bot>=20.0
email-validator>=2.0.0
slack-sdk>=3.20.0

# AI和机器学习（可选，较大）
# anthropic>=0.7.0
# openai>=1.0.0
# transformers>=4.30.0
# torch>=2.0.0
# tensorflow>=2.12.0
# xgboost>=1.7.0
# lightgbm>=3.3.0

# 监控和日志
prometheus-client>=0.16.0
loguru>=0.7.0

# 测试
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 工具
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0
colorama>=0.4.6